Remove get_loaders helpers from main
Refactors `get_loaders` and associated conditional compilation logic
from `main.rs` into `loaders/mock.rs` and `loaders/pgsqlx.rs` (see
7b527dc7af28d175bdd833223922e4a5144cf391 for background).
