+++ src/main.rs
@@ -1,32 +1,7 @@
 #![allow(where_clauses_object_safety)]
 
-#[cfg(not(feature = "mock"))]
-use sqlx::PgPool;
-
-#[cfg(feature = "mock")]
-use {
-    remote_server::database::{
-        mock,
-        schema::{
-            DatabaseRow, ItemLineRow, ItemRow, NameRow, RequisitionLineRow, RequisitionRow,
-            StoreRow, TransactLineRow, TransactRow, UserAccountRow,
-        },
-    },
-    std::collections::HashMap,
-};
-
 use remote_server::{
-    database::{
-        loader::{
-            ItemLineLoader, ItemLoader, NameLoader, RequisitionLineLoader, RequisitionLoader,
-            StoreLoader, TransactLineLoader, TransactLoader, UserAccountLoader,
-        },
-        repository::{
-            repository::get_repositories, ItemLineRepository, ItemRepository, NameRepository,
-            RequisitionLineRepository, RequisitionRepository, StoreRepository,
-            TransactLineRepository, TransactRepository, UserAccountRepository,
-        },
-    },
+    database::{loader::get_loaders, repository::get_repositories},
     server::{
         data::{ActorRegistry, LoaderMap, LoaderRegistry, RepositoryMap, RepositoryRegistry},
         middleware::{compress as compress_middleware, logger as logger_middleware},
@@ -40,196 +15,12 @@ use remote_server::{
 };
 
 use actix_web::{web::Data, App, HttpServer};
-use async_graphql::dataloader::DataLoader;
 use std::{
     env,
     net::TcpListener,
     sync::{Arc, Mutex},
     time::Duration,
 };
-
-#[cfg(not(feature = "mock"))]
-pub async fn get_loaders(settings: &Settings) -> LoaderMap {
-    let pool: PgPool = PgPool::connect(&settings.database.connection_string())
-        .await
-        .expect("Failed to connect to database");
-
-    let mut loaders: LoaderMap = LoaderMap::new();
-
-    let item_repository = ItemRepository::new(pool.clone());
-    let item_loader = DataLoader::new(ItemLoader { item_repository });
-
-    let item_line_repository = ItemLineRepository::new(pool.clone());
-    let item_line_loader = DataLoader::new(ItemLineLoader {
-        item_line_repository,
-    });
-
-    let requisition_repository = RequisitionRepository::new(pool.clone());
-    let requisition_loader = DataLoader::new(RequisitionLoader {
-        requisition_repository,
-    });
-
-    let requisition_line_repository = RequisitionLineRepository::new(pool.clone());
-    let requisition_line_loader = DataLoader::new(RequisitionLineLoader {
-        requisition_line_repository,
-    });
-
-    let name_repository = NameRepository::new(pool.clone());
-    let name_loader = DataLoader::new(NameLoader { name_repository });
-
-    let store_repository = StoreRepository::new(pool.clone());
-    let store_loader = DataLoader::new(StoreLoader { store_repository });
-
-    let transact_repository = TransactRepository::new(pool.clone());
-    let transact_loader = DataLoader::new(TransactLoader {
-        transact_repository,
-    });
-
-    let transact_line_repository = TransactLineRepository::new(pool.clone());
-    let transact_line_loader = DataLoader::new(TransactLineLoader {
-        transact_line_repository,
-    });
-
-    let user_account_repository = UserAccountRepository::new(pool.clone());
-    let user_account_loader = DataLoader::new(UserAccountLoader {
-        user_account_repository,
-    });
-
-    loaders.insert(item_loader);
-    loaders.insert(item_line_loader);
-    loaders.insert(requisition_loader);
-    loaders.insert(requisition_line_loader);
-    loaders.insert(name_loader);
-    loaders.insert(store_loader);
-    loaders.insert(transact_loader);
-    loaders.insert(transact_line_loader);
-    loaders.insert(user_account_loader);
-
-    loaders
-}
-
-#[cfg(feature = "mock")]
-pub async fn get_loaders(_settings: &Settings) -> LoaderMap {
-    let mut loaders = LoaderMap::new();
-
-    let mut mock_data: HashMap<String, DatabaseRow> = HashMap::new();
-
-    let mock_items: Vec<ItemRow> = mock::mock_items();
-    for item in mock_items {
-        mock_data.insert(item.id.to_string(), DatabaseRow::Item(item.clone()));
-    }
-
-    let mock_item_lines: Vec<ItemLineRow> = mock::mock_item_lines();
-    for item_line in mock_item_lines {
-        mock_data.insert(
-            item_line.id.to_string(),
-            DatabaseRow::ItemLine(item_line.clone()),
-        );
-    }
-
-    let mock_requisitions: Vec<RequisitionRow> = mock::mock_requisitions();
-    for requisition in mock_requisitions {
-        mock_data.insert(
-            requisition.id.to_string(),
-            DatabaseRow::Requisition(requisition.clone()),
-        );
-    }
-
-    let mock_requisition_lines: Vec<RequisitionLineRow> = mock::mock_requisition_lines();
-    for requisition_line in mock_requisition_lines {
-        mock_data.insert(
-            requisition_line.id.to_string(),
-            DatabaseRow::RequisitionLine(requisition_line.clone()),
-        );
-    }
-
-    let mock_names: Vec<NameRow> = mock::mock_names();
-    for name in mock_names {
-        mock_data.insert(name.id.to_string(), DatabaseRow::Name(name.clone()));
-    }
-
-    let mock_stores: Vec<StoreRow> = mock::mock_stores();
-    for store in mock_stores {
-        mock_data.insert(store.id.to_string(), DatabaseRow::Store(store.clone()));
-    }
-
-    let mock_transacts: Vec<TransactRow> = mock::mock_transacts();
-    for transact in mock_transacts {
-        mock_data.insert(
-            transact.id.to_string(),
-            DatabaseRow::Transact(transact.clone()),
-        );
-    }
-
-    let mock_transact_lines: Vec<TransactLineRow> = mock::mock_transact_lines();
-    for transact_line in mock_transact_lines {
-        mock_data.insert(
-            transact_line.id.to_string(),
-            DatabaseRow::TransactLine(transact_line.clone()),
-        );
-    }
-
-    let mock_user_accounts: Vec<UserAccountRow> = mock::mock_user_accounts();
-    for user_account in mock_user_accounts {
-        mock_data.insert(
-            user_account.id.to_string(),
-            DatabaseRow::UserAccount(user_account.clone()),
-        );
-    }
-
-    let mock_data: Arc<Mutex<HashMap<String, DatabaseRow>>> = Arc::new(Mutex::new(mock_data));
-
-    let item_repository = ItemRepository::new(Arc::clone(&mock_data));
-    let item_loader = DataLoader::new(ItemLoader { item_repository });
-
-    let item_line_repository = ItemLineRepository::new(Arc::clone(&mock_data));
-    let item_line_loader = DataLoader::new(ItemLineLoader {
-        item_line_repository,
-    });
-
-    let requisition_repository = RequisitionRepository::new(Arc::clone(&mock_data));
-    let requisition_loader = DataLoader::new(RequisitionLoader {
-        requisition_repository,
-    });
-
-    let requisition_line_repository = RequisitionLineRepository::new(Arc::clone(&mock_data));
-    let requisition_line_loader = DataLoader::new(RequisitionLineLoader {
-        requisition_line_repository,
-    });
-
-    let name_repository = NameRepository::new(Arc::clone(&mock_data));
-    let name_loader = DataLoader::new(NameLoader { name_repository });
-
-    let store_repository = StoreRepository::new(Arc::clone(&mock_data));
-    let store_loader = DataLoader::new(StoreLoader { store_repository });
-
-    let transact_repository = TransactRepository::new(Arc::clone(&mock_data));
-    let transact_loader = DataLoader::new(TransactLoader {
-        transact_repository,
-    });
-
-    let transact_line_repository = TransactLineRepository::new(Arc::clone(&mock_data));
-    let transact_line_loader = DataLoader::new(TransactLineLoader {
-        transact_line_repository,
-    });
-
-    let user_account_repository = UserAccountRepository::new(Arc::clone(&mock_data));
-    let user_account_loader = DataLoader::new(UserAccountLoader {
-        user_account_repository,
-    });
-
-    loaders.insert(item_loader);
-    loaders.insert(item_line_loader);
-    loaders.insert(requisition_loader);
-    loaders.insert(requisition_line_loader);
-    loaders.insert(name_loader);
-    loaders.insert(store_loader);
-    loaders.insert(transact_loader);
-    loaders.insert(transact_line_loader);
-    loaders.insert(user_account_loader);
-
-    loaders
-}
 
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
